feat: major fixes
